styling changes for Download page
